Minor bug corrections:
Readme: Removed the option par_redudancy. It was superseeded the the PAR_EXTRA_ARGS a while back. It was not doing anything.
uploadit: Added the switch -sfv (it didn't existed). Added the switch -nfo. The -nfo <path_to_nfo_file> will add a NFO file to the upload.
the goal of this new switch is to allow the nfo be on other folder and to upload it in the same level as the .rars
newsup: added the option to deal with status code 400