Added specific code to run the scripts in windows
The original function worked well in linux, but not so well in windows. So now, before running a script, the bot checks the platform in which is running (Linux or windows) and executes specific code for each of this platforms.
In linux we're using the exec call, while in windows we're using the backticks.

Removed some prints to stdout.