Reply to the initial ping command.
Some servers send a ping command immediatly after we connect. The bot original behaviour was to reply to the ping commands after the authentication was done. So if the first thing was sent before any authentication happened it would be impossible to proceed with the authentication because the IRC server wouldn't accept it. This would make the bot to be unable to connect.