Several changes affecting the read/write select, article groups, and reading from socket
Merged the different read write selects into one select.
Replaced the readline for a sysread - if we are using select, we should use sysread.
Changed the groups of an article from array ref to string.
Changed the return value of the function body from array ref to array, in order to minimize de-refs.